chore: Newest component split, Product styles changed
