+++ client/src/components/Home/Home.module.css
@@ -51,15 +51,3 @@
 .category-card:hover::before {
   background: rgba(0, 0, 0, 0.5);
 }
-.newest {
-  font-size: 1.8rem;
-  margin-block: 4rem;
-}
-.newest h3 {
-  margin-bottom: 1rem;
-}
-.newest-list {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
-}
