+++ client/src/components/Newest/Newest.module.css
@@ -1,0 +1,15 @@
+.newest {
+  font-size: 1.8rem;
+  margin-block: 4rem;
+}
+.newest img {
+  margin-bottom: 3rem;
+}
+.newest h3 {
+  margin-bottom: 1rem;
+}
+.newest-list {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
+}
